refactor(auth): Remove ktx imports
