+++ auth/app/src/main/java/com/google/firebase/quickstart/auth/kotlin/EmulatorSuite.kt
@@ -1,7 +1,7 @@
 package com.google.firebase.quickstart.auth.kotlin
 
-import com.google.firebase.auth.ktx.auth
-import com.google.firebase.ktx.Firebase
+import com.google.firebase.auth.auth
+import com.google.firebase.Firebase
 
 class EmulatorSuite {
