+++ auth/app/src/main/java/com/google/firebase/quickstart/auth/kotlin/MultiFactorActivity.kt
@@ -4,8 +4,8 @@ import android.app.Activity
 import android.os.Bundle
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.auth.FirebaseUser
-import com.google.firebase.auth.ktx.auth
-import com.google.firebase.ktx.Firebase
+import com.google.firebase.auth.auth
+import com.google.firebase.Firebase
 
 class MultiFactorActivity : Activity() {
     // [START declare_auth]
